% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_dic}
\alias{calculate_dic}
\title{Calculate DIC function.
Has two arguments: (1) log full likelihood at thetahat and (2) vector of log-likelihood at the theta samples
Calculate the DIC criteria values}
\usage{
calculate_dic(loglikatthetahat, logliks)
}
\arguments{
\item{loglikatthetahat}{Log of the likelihood function at theta hat (Bayes). It is a scalar value.}

\item{logliks}{A vector of log likelihood values at the theta samples}
}
\value{
a list containing four values pdic, pdicalt, dic and dicalt
}
\description{
Calculate DIC function.
Has two arguments: (1) log full likelihood at thetahat and (2) vector of log-likelihood at the theta samples
Calculate the DIC criteria values
}
