% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_waic}
\alias{calculate_waic}
\title{Calculate WAIC  function.
Has the sole argument component wise likelihood at thetahat samples.
v must be a matrix
Calculate the WAIC criteria values}
\usage{
calculate_waic(v)
}
\arguments{
\item{v}{must be a N (MCMC) by n (data) sample size matrix
of the log-likelihood evaluations}
}
\value{
a list containing four values p_waic, p_waic alt, waic and waic_alt
}
\description{
Calculate WAIC  function.
Has the sole argument component wise likelihood at thetahat samples.
v must be a matrix
Calculate the WAIC criteria values
}
