% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.bmstdr}
\alias{coef.bmstdr}
\title{Prints and returns the estimates of the coefficients}
\usage{
\method{coef}{bmstdr}(object, digits = 3, ...)
}
\arguments{
\item{object}{A bmstdr model fit object.}

\item{digits}{How many significant digits after the decimal to print, defaults to 3.}

\item{...}{Any other additional arguments}
}
\value{
Coefficients are returned as a data frame preserving the names of 
the covariates
}
\description{
Prints and returns the estimates of the coefficients
}
