% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_data.R
\docType{data}
\name{colpalette}
\alias{colpalette}
\title{The color palette used to draw maps to illustrate the package 
bmstdr, see  \insertCite{Sahubook;textual}{bmstdr}
It has the values in order: dodgerblue4, dodgerblue2, firebrick2,  
firebrick4 and purple.}
\format{
An object of class \code{character} of length 5.
}
\usage{
colpalette
}
\description{
The color palette used to draw maps to illustrate the package 
bmstdr, see  \insertCite{Sahubook;textual}{bmstdr}
It has the values in order: dodgerblue4, dodgerblue2, firebrick2,  
firebrick4 and purple.
}
\examples{
 colpalette
}
\references{
\insertAllCited{}
}
\keyword{datasets}
