% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunctions.R
\name{fig11.13.plot}
\alias{fig11.13.plot}
\title{Draws a time series (ribbon) plot by combining fitted and predicted values}
\usage{
fig11.13.plot(yobs, ylow, ymed, yup, misst)
}
\arguments{
\item{yobs}{A vector of the observed values}

\item{ylow}{A vector of the lower limits of the fitted or predicted values}

\item{ymed}{A vector of fitted or predicted values}

\item{yup}{A vector of the upper limits of the fitted or predicted values}

\item{misst}{An integer vector which contains the indices of the predicted
values}
}
\value{
A ribbon plot,  ggplot2 object,  which shows observed values
in red color and open circle, predicted values in blue color and
filled circle.
}
\description{
Draws a time series (ribbon) plot by combining fitted and predicted values
}
