% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.bmstdr}
\alias{fitted.bmstdr}
\title{Extract fitted values from bmstdr objects.}
\usage{
\method{fitted}{bmstdr}(object, ...)
}
\arguments{
\item{object}{A bmstdr model fit object.}

\item{...}{Any other additional arguments.}
}
\description{
Extract fitted values from bmstdr objects.
}
