% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_data.R
\docType{data}
\name{gridnysptime}
\alias{gridnysptime}
\title{Values of three covariates for 100 grid locations in New York for
the 62 days during the months of July and August, 2006.}
\format{
An object of class \code{data.frame} with 6200 rows and 11 columns.

An object of class \code{data.frame} with 6200 rows and 11 columns.
}
\source{
It is the same data set as NYgrid data set in the spTimer package, 
\insertCite{spTimer;textual}{bmstdr}.
with two added columns providing the UTM X- and Y- coordinates.
 Each data row is  for a particular grid site and day as detailed below.

 @format A data frame with 6200 rows and 11 columns:
\describe{
  \item{s.index}{site index (1 to 100)}
  \item{Longitude}{Longitude of the site}
  \item{Latitude}{Latitude of the site}
  \item{utmx}{UTM X-coordinate of the site}
  \item{utmy}{UTM Y-coordinate of the site}
  \item{Year}{This is 2006 for all the rows}
  \item{Month}{Month taking values 7 for July and 8 for August}
  \item{Day}{Day taking values 1 to 31}
  \item{xmaxtemp}{Maximum temperature (degree Celsius)}
  \item{xwdsp}{wind speed (nautical mile per hour)}
  \item{xrh}{Relative humidity}
}

It is the same data set as NYgrid data set in the spTimer package, 

with two added columns providing the UTM X- and Y- coordinates.
 Each data row is  for a particular grid site and day as detailed below.

 @format A data frame with 6200 rows and 11 columns:
\describe{
  \item{s.index}{site index (1 to 100)}
  \item{Longitude}{Longitude of the site}
  \item{Latitude}{Latitude of the site}
  \item{utmx}{UTM X-coordinate of the site}
  \item{utmy}{UTM Y-coordinate of the site}
  \item{Year}{This is 2006 for all the rows}
  \item{Month}{Month taking values 7 for July and 8 for August}
  \item{Day}{Day taking values 1 to 31}
  \item{xmaxtemp}{Maximum temperature (degree Celsius)}
  \item{xwdsp}{wind speed (nautical mile per hour)}
  \item{xrh}{Relative humidity}
}
}
\usage{
gridnysptime

gridnysptime
}
\description{
Values of three covariates for 100 grid locations in New York for
the 62 days during the months of July and August, 2006.

Values of three covariates for 100 grid locations in New York for
the 62 days during the months of July and August, 2006.
}
\examples{
 head(gridnysptime)
 summary(gridnysptime[, 9:11])
 summary(gridnysptime[, 9:11])
}
\references{
\insertAllCited{}

\insertAllCited{}
}
\keyword{datasets}
