% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunctions.R
\name{hitandfalsealarm}
\alias{hitandfalsealarm}
\title{Calculate the hit and false alarm rates}
\usage{
hitandfalsealarm(thresh, yobs, ypred)
}
\arguments{
\item{thresh}{Threshold value}

\item{yobs}{A vector of observations, may include missing values}

\item{ypred}{Predictions}
}
\value{
A list containing the calculated hit and false alarm rates
}
\description{
Calculate the hit and false alarm rates
}
