% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.bmstdr}
\alias{is.bmstdr}
\title{Is it a bmstdr model fitted object?}
\usage{
is.bmstdr(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A TRUE/FALSE logical output.
}
\description{
Is it a bmstdr model fitted object?
}
