% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunctions.R
\name{materncov}
\alias{materncov}
\title{Banerjee, Carlin and Gelfand (2015) Mat'ern covariance function}
\usage{
materncov(d, phi, nu)
}
\arguments{
\item{d}{is the distance}

\item{phi}{is the rate of decay}

\item{nu}{is the smoothness parameter}
}
\value{
Returns the Mat'ern covariance for distance object d
}
\description{
Banerjee, Carlin and Gelfand (2015) Mat'ern covariance function
}
