% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunctions.R
\name{maternfun}
\alias{maternfun}
\title{Banerjee et al Mat'ern covariance function}
\usage{
maternfun(d, phi, nu)
}
\arguments{
\item{d}{is the distance}

\item{phi}{is the rate of decay}

\item{nu}{is the smoothness parameter}
}
\value{
Returns the Mat'ern correlation function for distance object d
}
\description{
Banerjee et al Mat'ern covariance function
}
