% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{phichoice_sp}
\alias{phichoice_sp}
\title{Calculates the validation statistics using the spatial model with a given range of values of
the decay parameter phi.}
\usage{
phichoice_sp(
  phis = seq(from = 0.1, to = 1, by = 0.1),
  s = c(8, 11, 12, 14, 18, 21, 24, 28),
  ...
)
}
\arguments{
\item{phis}{A vector values of phi}

\item{s}{A vector giving the validation sites}

\item{...}{Any additional parameter that may be passed to \code{Bspatial}}
}
\value{
A data frame giving the phi values and the corresponding validation statistics
}
\description{
Calculates the validation statistics using the spatial model with a given range of values of
the decay parameter phi.
}
