% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{phichoicep}
\alias{phichoicep}
\title{Calculates the validation statistics using the spatial model with a given range of values of
the decay parameter phi.}
\usage{
phichoicep(
  phis = c(0.001, 0.005, 0.025, 0.125, 0.625),
  phit = c(0.05, 0.25, 1.25, 6.25),
  valids = c(8, 11, 12, 14, 18, 21, 24, 28),
  verbose = TRUE
)
}
\arguments{
\item{phis}{A vector values of phi for spatial decay}

\item{phit}{A vector values of phi for temporal decay}

\item{valids}{A vector giving the validation sites}

\item{verbose}{Logical. Should progress be printed?}
}
\value{
A data frame giving the phi values and the corresponding validation statistics
}
\description{
Calculates the validation statistics using the spatial model with a given range of values of
the decay parameter phi.
}
\examples{
\donttest{
asave <-  phichoicep()
}
}
