% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bmstdr}
\alias{print.bmstdr}
\title{Provides basic information regarding the fitted model.}
\usage{
\method{print}{bmstdr}(x, digits = 3, ...)
}
\arguments{
\item{x}{A bmstdr model fit object.}

\item{digits}{How many significant digits after the decimal to print, defaults to 3.}

\item{...}{Any other additional arguments}
}
\value{
No return value, called for side effects.
}
\description{
Provides basic information regarding the fitted model.
}
