% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{resid.bmstdr}
\alias{resid.bmstdr}
\title{Extract residuals from a bmstdr  fitted object.}
\usage{
\method{resid}{bmstdr}(object, ...)
}
\arguments{
\item{object}{A bmstdr model fit object.}

\item{...}{Any other additional arguments.}
}
\description{
Extract residuals from a bmstdr  fitted object.
}
