% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.bmstdr}
\alias{residuals.bmstdr}
\title{Extract residuals from a bmstdr  fitted object.}
\usage{
\method{residuals}{bmstdr}(object, numbers = NULL, ...)
}
\arguments{
\item{object}{A bmstdr model fit object.}

\item{numbers}{a list with two components: sn=number of spatial locations 
tn=number of time points. Residuals will be assumed to follow the arrangement 
of the data frame - sorted by space and then time within space.}

\item{...}{Any other additional arguments.}
}
\value{
Returns a vector of residuals. If appropriate, it draws a 
time series plot of residuals. Otherwise, it draws a plot of residuals 
against observation numbers.
}
\description{
Extract residuals from a bmstdr  fitted object.
}
