% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.bmstdr}
\alias{summary.bmstdr}
\title{Provides basic summaries of model fitting.}
\usage{
\method{summary}{bmstdr}(object, digits = 3, ...)
}
\arguments{
\item{object}{A bmstdr model fit object.}

\item{digits}{How many significant digits after the decimal to print, defaults to 3.}

\item{...}{Any other additional arguments}
}
\value{
No return value, called for side effects.
}
\description{
Provides basic summaries of model fitting.
}
