% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ydata.R
\docType{data}
\name{ydata}
\alias{ydata}
\title{Average air pollution values from 28 sites in New York.}
\format{
A vector with 28 real values.

An object of class \code{numeric} of length 28.
}
\source{
See the NYdata set in spTimer package.
 Each data point is the mean of the available daily 8-hour maximum average ozone concentrations
 in parts per billion (ppb) at each of the
 28 sites. The daily values are for the month of July and August in 2006.
}
\usage{
ydata

ydata
}
\description{
Average air pollution values from 28 sites in New York.
}
\examples{
\dontrun{
 ydata
}
}
\keyword{datasets}
