% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{terms.bmstdr}
\alias{terms.bmstdr}
\title{Prints the terms}
\usage{
\method{terms}{bmstdr}(x, ...)
}
\arguments{
\item{x}{A bmstdr model fit object.}

\item{...}{Any other additional arguments}
}
\value{
Terms in the model formula
}
\description{
Prints the terms
}
