% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_data.R
\docType{data}
\name{ydata}
\alias{ydata}
\title{Average air pollution values from 28 sites in New York.}
\format{
A vector with 28 real values.

An object of class \code{numeric} of length 28.
}
\source{
This is obtained by calculating site-wise averages 
of the NYdata set in the 'spTimer' package
\insertCite{spTimer;textual}{bmstdr}.
 Each data point is the mean of the available daily 8-hour maximum average ozone concentrations
 in parts per billion (ppb) at each of the
 28 sites. The daily values are for the month of July and August in 2006.
}
\usage{
ydata

ydata
}
\description{
Average air pollution values from 28 sites in New York.
}
\examples{
 summary(ydata)
}
\references{
\insertAllCited{}
}
\keyword{datasets}
